fix/import CreditHistoryListContainerとCreditHistoryListPresenterを削除し、CreditHistoryTableコンポーネントを新規作成 ・`CreditHistoryListContainer.tsx`および`CreditHistoryListPresenter.tsx`を削除した。 ・代わりに、`DataTable`を使用した`CreditHistoryTable.tsx`コンポーネントを新規作成し、引落日やクレジットカード情報、アカウント、金額を効率的に表示できるようにした。 ・リファクタリングによりコードの簡素化と保守性向上を目的とした対応である。
