+++ apps/web/src/features/householdCreditHistoryList/components/credit-history-table.tsx
@@ -1,0 +1,63 @@
+"use client";
+
+import type { YYYY_MM_DD } from "@/util/date/date";
+import { useRouter } from "next/navigation";
+import type { FC } from "react";
+
+import { FormatPrice } from "../../../components/molecules/FormatPrice";
+import { DataTable } from "../../../components/ui/v4/table";
+import { IocomeType } from "../../../domain/model/household/IocomeType";
+
+type Props = {
+  creditHistoryList: {
+    id: string;
+    withdrawalDate: YYYY_MM_DD;
+    creditCard: string;
+    accountName: string;
+    totalAmount: number;
+  }[];
+};
+
+export const CreditHistoryTable: FC<Props> = ({ creditHistoryList }) => {
+  const { push } = useRouter();
+
+  const showDetailPage = (summaryId: string) => {
+    push(`/household/creditCard/${summaryId}`);
+  };
+
+  return (
+    <DataTable
+      columns={[
+        {
+          accessor: "withdrawalDate",
+          title: "引落日",
+          width: "50",
+        },
+        {
+          accessor: "creditCard",
+          title: "クレジットカード",
+          width: "100",
+        },
+        {
+          accessor: "accountName",
+          title: "アカウント",
+          width: "100",
+        },
+        {
+          accessor: "totalAmount",
+          title: "合計金額",
+          render: (v) => {
+            return (
+              <FormatPrice
+                price={v.totalAmount as number}
+                iocomeType={IocomeType.Outcome}
+              />
+            );
+          },
+        },
+      ]}
+      records={creditHistoryList}
+      onRowClick={(v) => showDetailPage(v.id)}
+    />
+  );
+};
